% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POST_label_switch.R
\name{align_MC_samples}
\alias{align_MC_samples}
\title{Reorder cluster labels after correction}
\usage{
align_MC_samples(
  res,
  cluster_record,
  state_record,
  Q_record = NULL,
  Pi_record = NULL
)
}
\arguments{
\item{res}{output from \code{ECR()}.}

\item{cluster_record}{same input as in \code{ECR()}.}

\item{state_record}{A T by K by U array of clonal copy number states, where T is the number of MCMC iterations, K is the number of clusters and U is the number of genes.}

\item{Q_record}{A list of T MCMC samples of transition matrices, each sample contains K S by S matrices, where K is the number of clusters and S is the number of states.}

\item{Pi_record}{State probability in the same format as \code{Q_record}.}
}
\description{
Align the cluster labels as well as all cluster-specific parameters
according to the label order returned by \code{ECR()}
for all MCMC samples.
}
\keyword{internal}
