% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POST_label_switch.R
\name{ECR}
\alias{ECR}
\title{Label Switching correction}
\usage{
ECR(cluster_record, loglik_record)
}
\arguments{
\item{cluster_record}{T by N matrix of MCMC clustering samples,
where T is the total number of Gibbs iteration and N is the number of cells being clustered.}

\item{loglik_record}{likelihood of each Gibbs iteration.}
}
\value{
The order of cluster labels in each MCMC iteration.
}
\description{
Wrapper of the Equivalence Classes Representatives (ECR) algorithm in \href{https://cran.r-project.org/web/packages/label.switching/index.html}{\code{label.switching}}
}
\keyword{internal}
