% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIM_details.R
\name{generate_cluster_states}
\alias{generate_cluster_states}
\title{Simulate cell clustering and state profile for each cluster}
\usage{
generate_cluster_states(K, U, Q, neutral_idx, overlap = TRUE)
}
\description{
Simulate cell clustering and state profile for each cluster
}
\keyword{internal}
