% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIM_details.R
\name{generate_states}
\alias{generate_states}
\title{Simulate state profiles for a clone.}
\usage{
generate_states(Q, U, neutral_idx)
}
\description{
Simulate state profiles for a clone.
}
\keyword{internal}
