% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PREP_prior.R
\name{get_priors}
\alias{get_priors}
\title{Generate prior parameters}
\usage{
get_priors(
  K,
  S,
  signal_RDR,
  signal_BAF,
  mu = log2(c(0.5, 1, 1.5, 2)),
  sigma_shape = 10,
  beta_shape1 = 1,
  beta_shape2 = 5,
  Q_diag = 10,
  neutral_idx = 2,
  to_neutral = 1.5
)
}
\arguments{
\item{K}{The maximum number of clusters.}

\item{S}{The number of states.}

\item{signal_RDR}{TRUE if prior for mu and sigma is required.}

\item{signal_BAF}{TRUE if prior for beta is required.}

\item{mu}{The prior distribution of mean of RDR for each state is N(\code{mu}, 0.1).}

\item{sigma_shape}{The prior distribution of the stand error of RDR for each state is Inv-Gamma(\code{sigma_shape}, 1).}

\item{beta_shape1}{The prior distribution of BAF related parameter is Beta(\code{beta_shape1}, \code{beta_shape2}).}

\item{beta_shape2}{See \code{beta_shape1}}

\item{Q_diag}{The prior of transition probability is a \code{S}\eqn{\times}\code{S} matrix, with off-diagonal entries 1 and diagonal elements \code{Q_diag}.}

\item{neutral_idx}{The index of the neutral state.}

\item{to_neutral}{Entries in the prior of transition probability that represent P(neutral state | non-neutral state).}
}
\value{
A list of prior parameters for MCMC sampling.
}
\description{
Generate prior parameters
}
\keyword{internal}
