% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIM_main.R
\name{get_sim_data}
\alias{get_sim_data}
\title{Clonal CNA simulation}
\usage{
get_sim_data(
  K,
  N,
  U,
  S = 4,
  expr = TRUE,
  RDR = TRUE,
  BAF = TRUE,
  RDR_var = NULL,
  norm_clone_prob = 0.2,
  CNV_overlap = TRUE,
  RDR_outlier_cnt = 0,
  BAF_missing_percent = 0.9,
  BAF_Dmax = 10,
  Q_to_neutral = 2
)
}
\arguments{
\item{K}{The number of clusters.}

\item{N}{The number of cells.}

\item{U}{The number of genes.}

\item{expr}{TRUE if gene count is to be generated.}

\item{RDR}{TRUE if RDR is to be generated. If expr == TRUE, RDR will be obtained from expr, otherwise generated from Normal distribution.}

\item{BAF}{TRUE if BAF is to be generated.}

\item{RDR_var}{The variance of RDR if generated from Normal distribution, needed iff expr == FALSE and RDR == TRUE.}

\item{norm_clone_prob}{The proportion of normal cluster among all cells.}

\item{CNV_overlap}{TRUE if each cluster has an identical interval with another cluster, which increases the similarity between different clusters.}

\item{RDR_outlier_cnt}{An integer. The number of outlier cells with random CN profiles.}

\item{BAF_missing_percent}{The percentage of NA in BAF data, effective only if BAF == TRUE.}

\item{BAF_Dmax}{The largest number of total count in BAF. The smaller it is the weaker the BAF signal is.}
}
\value{
Simulated data as well as cell cluster labels and CN state profile for each cluster
}
\description{
Create scRNA-seq count data, RDR or BAF for testing \code{Chloris}
}
\examples{
sims <- get_sim_data(K = 4, N = 100, U = 200)
names(sims)
plot_inout(sims$RDR, list(sims$cluster_true), sims$states_true, state_mean = c(-1, 0, 0.5, 1.5))

}
