% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIM_details.R
\name{get_true_transition}
\alias{get_true_transition}
\title{Simulate transition matrix}
\usage{
get_true_transition(
  K,
  S,
  diag,
  to_neutral,
  neutral_idx,
  off_diag_variance = 0.1
)
}
\description{
Simulate transition matrix
}
\keyword{internal}
