% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{preprocess}
\alias{preprocess}
\title{Basic preprocess}
\usage{
preprocess(
  counts,
  ref_cells,
  window_smooth = NULL,
  recentering = FALSE,
  psudo_count = 0
)
}
\arguments{
\item{counts}{Raw gene count matrix, gene*cell}

\item{ref_cells}{A vector of reference or normal cell index.}

\item{window_smooth}{window size for window smoothing, NULL then no window smoothing applied}

\item{recentering}{Recenter RDR for each cell independently.}

\item{psudo_count}{Psudo count in log2 transform}
}
\description{
Basic preprocess
}
