% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIM_details.R
\name{preprocess_sims}
\alias{preprocess_sims}
\title{Preprocess simulated gene expression data to RDR and summary(RDR|state)}
\usage{
preprocess_sims(count_data, cell_level_RDR_states)
}
\description{
Preprocess simulated gene expression data to RDR and summary(RDR|state)
}
\keyword{internal}
