% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIM_details.R
\name{sim_expr_Splat}
\alias{sim_expr_Splat}
\title{Simulate gene count}
\usage{
sim_expr_Splat(cell_level_states, CN_levels, outlier_cnt = 0, verbose = FALSE)
}
\description{
Simulate gene count
}
\keyword{internal}
