% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIM_details.R
\name{sim_RDR}
\alias{sim_RDR}
\title{Simulate RDR}
\usage{
sim_RDR(
  cell_level_states,
  CN_levels,
  from_Splat,
  var_true = NULL,
  RDR_outlier_cnt = 0
)
}
\description{
Simulate RDR
}
\keyword{internal}
