% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIM_details.R
\name{Splat_GeneMeans}
\alias{Splat_GeneMeans}
\title{Simulate gene means from Gamma distribution for gene count}
\usage{
Splat_GeneMeans(U)
}
\description{
Simulate gene means from Gamma distribution for gene count
}
\keyword{internal}
