% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIM_details.R
\name{Splat_meanvar}
\alias{Splat_meanvar}
\title{Simulate mean-variance trend for gene count}
\usage{
Splat_meanvar(BaseMeans, outlier_cnt = 0)
}
\description{
Simulate mean-variance trend for gene count
}
\keyword{internal}
