% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLOT_inout.R
\name{plot_inout}
\alias{plot_inout}
\title{Main plot function}
\usage{
plot_inout(
  input,
  type = "RDR",
  cluster_labels = NULL,
  CN_states = NULL,
  state_mean = NULL,
  lim = NULL,
  break_idx = NULL,
  cluster_colour = NULL
)
}
\arguments{
\item{input}{Matrix of RDR or BAF, where each row is a gene and each column represents a cell.}

\item{type}{"RDR" or "BAF".}

\item{cluster_labels}{A list of one or more cell partitions, each partition is represented as
an N dimensional vector of cluster labels, N is the number of cells.
The columns in \code{input} will be ordered according to the first element in \code{cluster_labels}.
If \code{cluster_labels == NULL}, the order of columns in \code{input} will be maintained.}

\item{CN_states}{A K by U matrix of copy number states in K clusters with U genes.}

\item{state_mean}{Only needed if \code{!is.null(CN_states)}.
The \code{CN_states} will be plotted together with \code{input} as one heatmap, where each state is
represented by the according value in \code{state_mean}. So the main purpose of setting \code{state_mean}
is to control the colour representations of \code{CN_states}.
If \code{state_mean == NULL}, the default value is \code{mean(input|s)}, where s is one of the states.}

\item{lim}{The imposed range of \code{input}.}

\item{break_idx}{The index of the first gene in each chromosome. This will generate a horizontal line for each index.}

\item{cluster_colour}{A vector of colours for each cluster.}
}
\value{
gg object
}
\description{
Plot input RDR and output clustering and clonal copy number profiles.
}
\examples{
\dontrun{
sims <- get_sim_data(K = 4, N = 100, U = 200)
plot_inout(sims$RDR)
plot_inout(sims$RDR, list(sims$cluster_true), sims$states_true, state_mean = c(-1, 0, 0.5, 1.5))
}
}
